move test helpers into suite file
