+++ metadata/metadata_test.go
@@ -3,51 +3,14 @@ package metadata_test
 import (
 	. "github.com/18F/cf-gitaware/metadata"
 
-	"github.com/codeskyblue/go-sh"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	"io/ioutil"
 )
-
-func createTempDir() (string, error) {
-	return ioutil.TempDir("", "metadata_test")
-}
-
-func createEmptyRepo(path string) error {
-	return sh.Command("git", "init", path).Run()
-}
-
-func createInitialCommit(repoPath string) error {
-	return sh.Command(
-		"git",
-		"commit",
-		"--allow-empty",
-		"-m", "\"test commit\"",
-		sh.Dir(repoPath),
-	).Run()
-}
-
-func createTestRepo() (string, error) {
-	path, err := createTempDir()
-	if err != nil {
-		return "", err
-	}
-	err = createEmptyRepo(path)
-	if err != nil {
-		return "", err
-	}
-	err = createInitialCommit(path)
-	if err != nil {
-		return "", err
-	}
-	return path, nil
-}
 
 var _ = Describe("Metadata", func() {
 	Describe(".GetRevision()", func() {
 		It("returns a 41-character string", func() {
-			repoPath, err := createTestRepo()
+			repoPath, err := CreateTestRepo()
 			Expect(err).NotTo(HaveOccurred())
 
 			rev, err := GetRevision(repoPath)
@@ -56,7 +19,7 @@ var _ = Describe("Metadata", func() {
 		})
 
 		It("returns an error when the directory isn't a repository", func() {
-			tempDirName, err := createTempDir()
+			tempDirName, err := CreateTempDir()
 			Expect(err).NotTo(HaveOccurred())
 
 			_, err = GetRevision(tempDirName)
